+++ examples/QuickPID_Basic/QuickPID_Basic.ino
@@ -9,7 +9,7 @@
 #define PIN_OUTPUT 3
 
 //Define Variables we'll be connecting to
-int Setpoint, Input, Output;
+float Setpoint, Input, Output;
 
 //Specify the links and initial tuning parameters
 float Kp = 2, Ki = 5, Kd = 1;
